+++ packages/expo-updates/plugin/src/withUpdates.ts
@@ -1,3 +1,4 @@
+import { getAccountUsername } from '@expo/config';
 import { ConfigPlugin, createRunOncePlugin } from '@expo/config-plugins';
 
 import { withUpdatesAndroid } from './withUpdatesAndroid';
@@ -6,8 +7,8 @@ import { withUpdatesIOS } from './withUpdatesIOS';
 const pkg = require('expo-updates/package.json');
 
 const withUpdates: ConfigPlugin<{ expoUsername?: string } | void> = (config, props = {}) => {
-  // The username should be passed from the CLI when the plugin is automatically used.
-  const expoUsername = (props || {}).expoUsername ?? process.env.EXPO_CLI_USERNAME ?? null;
+  // The username will be passed from the CLI when the plugin is automatically used.
+  const expoUsername = (props || {}).expoUsername ?? getAccountUsername(config);
 
   config = withUpdatesAndroid(config, { expoUsername });
   config = withUpdatesIOS(config, { expoUsername });